+++ plugins/install-clash-cli.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+source "$HOME/.clash-cli/bootstrap.sh"
+
+URL_CLASH_CLI_REPO="https://github.com/BlueSky-07/clash-cli"
+
+dependency_check git "https://git-scm.com/downloads/"